modified:   src/lib.rs

+++ src/lib.rs
@@ -33,9 +33,8 @@ mod tests {
         let mut buffer: Vec<u8> = vec![0; bufsize];
 
         /* compress to memory */
-        let fpz = unsafe {
-            fpzip_write_to_buffer(buffer.as_mut_ptr() as *mut std::ffi::c_void, bufsize as u64)
-        };
+        let fpz =
+            unsafe { fpzip_write_to_buffer(buffer.as_mut_ptr() as *mut std::ffi::c_void, bufsize) };
 
         unsafe {
             (*fpz).type_ = FPZIP_TYPE_FLOAT as i32;
